Migrate composer -> apacheairflow
Support a log format "Marking task as ..."
